course.anchor: removes teacher_address field
Now that we'll have DNA-per-course, teacher_address field would be the same for all anchors in any course DNA.
Because only DNA creator (==teacher) is capable of creating new courses.
Considering that originally this field is there to increase uniqueness of the anchor, it's no longer needed.
Instead this commit adds field course_title which would have high probability to be unique across any given DNA
(just because it's quite unlikely for a teacher to create 2 courses with identical titles inside a single DNA).